Split styles and scripts into separate files - 2
