+++ src/common.blocks/skills-button.scss
@@ -1,0 +1,9 @@
+.skills-button {
+  border-radius: 50%;
+  border-width: 0;
+  color: var(--background);
+  background-color: var(--foreground);
+  font-size: 1.2em;
+  width: 25px;
+  height: 25px;
+}
