+++ src/common.blocks/skills-list/skills-list.scss
@@ -1,0 +1,12 @@
+.skills-list {
+  display: flex;
+  gap: 10px;
+  flex-direction: column;
+
+  &:empty{
+    &::after {
+      margin: 10px auto;
+      content: 'No skills yet. Feel free to add them ↓';
+    }
+  }
+}
