+++ src/skills.html
@@ -28,7 +28,7 @@
             </label>
             <div class="skill__progress-bar progress-bar"></div>
             <div class="skill__remove-button-container">
-                <button type="submit" class="add-button focusable"><i class="far fa-plus" title="Add skill"></i></button>
+                <button type="submit" class="skills-button focusable"><i class="far fa-plus" title="Add skill"></i></button>
             </div>
         </div>
     </form>
