+++ src/styles/main.scss
@@ -6,6 +6,7 @@ $fa-font-path: "./assets/webfonts";
 @import "~@fortawesome/fontawesome-free/scss/fontawesome.scss";
 @import "~@fortawesome/fontawesome-free/scss/regular.scss";
 
+// index.html
 @import "../common.blocks/light-theme";
 @import "../common.blocks/link";
 @import "../common.blocks/focusable";
@@ -17,6 +18,13 @@ $fa-font-path: "./assets/webfonts";
 @import "../common.blocks/minor-info";
 @import "../common.blocks/social";
 @import "../common.blocks/social__icon";
+
+// skills.html
+@import "../common.blocks/progress-bar";
+@import "../common.blocks/skill/skill";
+@import "../common.blocks/skills-button";
+@import "../common.blocks/skills-list";
+@import "../common.blocks/skills-list-container";
 
 // default theme
 body {
@@ -38,138 +46,4 @@ html {
 
 html, body {
   margin: 0;
-}
-
-.skills-list-container {
-  width: 400px;
-  margin: 0 auto;
-  background-color: var(--background);
-  color: var(--foreground);
-  font-size: 14px;
-  padding: 10px 10px 15px;
-  border-radius: 6px;
-
-  display: flex;
-  gap: 10px;
-  flex-direction: column;
-
-  &__form {
-    margin: 5px -10px -10px;
-    padding: 10px;
-    border-radius: inherit;
-  }
-}
-
-.skills-list {
-  display: flex;
-  gap: 10px;
-  flex-direction: column;
-
-  &:empty{
-    &::after {
-      margin: 10px auto;
-      content: 'No skills yet. Feel free to add them ↓';
-    }
-  }
-}
-
-.skill {
-  display: grid;
-  grid-template-columns: 1fr auto auto;
-  grid-column-gap: 10px;
-  grid-row-gap: 5px;
-
-  &__name {
-    grid-column: 1;
-    grid-row: 1;
-
-    font-weight: bold;
-
-    &::after {
-      content: var(--skill-name);
-    }
-
-    &_in-form::after {
-      content: '';
-    }
-  }
-
-  &__value {
-    grid-column: 2;
-    grid-row: 1;
-
-    opacity: 0.65;
-
-    &::after {
-      content: var(--skill-value);
-    }
-
-    &_in-form::after {
-      content: '';
-    }
-  }
-
-  &__progress-bar {
-    grid-column-start: 1;
-    grid-column-end: 3;
-    grid-row: 2;
-  }
-
-  &__remove-button-container {
-    grid-column: 3;
-    grid-row-start: 1;
-    grid-row-end: 3;
-
-    display: flex;
-    align-items: end;
-  }
-}
-
-.remove-button, .add-button {
-  border-radius: 50%;
-  border-width: 0;
-  color: var(--background);
-  background-color: var(--foreground);
-  font-size: 1.2em;
-  width: 25px;
-  height: 25px;
-}
-
-.progress-bar {
-  background-position: left;
-  background-origin: content-box;
-  background-clip: content-box;
-  background-image: linear-gradient(90deg, var(--skill-accent) var(--skill-value), transparent var(--skill-value));
-  box-sizing: border-box;
-  height: 15px;
-  border-radius: 5px;
-  padding: 2px;
-  border: 2px solid var(--skill-accent);
-}
-
-.skill_form {
-  background-color: var(--accent-background);
-
-  &__title {
-    font-size: 1.5em;
-    margin-bottom: 10px;
-    margin-top: 5px;
-  }
-
-  &__input {
-    color: inherit;
-    background-color: inherit;
-    font: inherit;
-    border-width: 1px;
-
-    &_name {
-      display: inline-block;
-      width: 190px;
-      font-weight: normal;
-    }
-  }
-}
-
-.skill__value_in-form {
-  opacity: 1;
 }
